Specify argument type in #[init] annotation

+++ font-tables/src/tables/hmtx.rs
@@ -6,7 +6,7 @@ pub const TAG: Tag = Tag::new(b"hmtx");
 
 font_types::tables! {
     /// The [hmtx (Horizontal Metrics)](https://docs.microsoft.com/en-us/typography/opentype/spec/hmtx) table
-    #[init(number_of_h_metrics, num_glyphs)]
+    #[init(number_of_h_metrics = "u16", num_glyphs = "u16")]
     Hmtx<'a> {
         /// Paired advance width and left side bearing values for each
         /// glyph. Records are indexed by glyph ID.
@@ -26,6 +26,6 @@ font_types::tables! {
     }
 }
 
-fn n_glyphs_less_n_metrics(num_glyphs: usize, num_metrics: usize) -> usize {
-    num_glyphs.saturating_sub(num_metrics)
+fn n_glyphs_less_n_metrics(num_glyphs: u16, num_metrics: u16) -> usize {
+    num_glyphs.saturating_sub(num_metrics) as usize
 }
